requirements.txt
aiohttp==3.10.11
beautifulsoup4==4.12.2
python-telegram-bot==20.7
lxml==4.9.3
requests==2.32.4